requirements.txt
SpeechRecognition
google-cloud-speech
oauth2client
simpleaudio
pyaudio
google-auth